AUI-1238 Migrate Modal markup to Bootstrap 3
Bootstrap 3's modal structure has 3 containers: .modal, .modal-dialog and .modal-content. The '.modal' container is not important for AUI though, and it would make the integration a lot more complicated, so I've just removed it and I'm making the widget work with .modal-dialog and .modal-content as the bounding box and the content box respectively.

With this the modal's features work as expected. There are just a few more Bootstrap css changes that we may want to incorporate later, but decided to leave out for now as users can do that in their code if necessary. The changes are:
(1) .modal_content doesn't have the 'overflow: hidden' rule anymore, so resizing to make the modal very small will show contents outside the modal.
(2) .modal-dialog has 'margin: 30px auto' by default, so resizing will grow both sides of the modal at the same time now.
(3) .btn-toolbar now clears float by default, so the close button takes more space than before, pushing the header content down.

Regarding (1), another thing we can do is allow the user to specify min width/height through ResizeConstrained plugin. We need to change AUI's modal to allow for this yet though, since ResizeConstrained needs to be plugged in after the Resize plugin and the modal isn't doing that yet.