AUI-1254 Fix bootstrap 3 icons on the page showing up on top of datepicker's overlay
The datepicker seems to actually be working by design. This styling issue is happening in our demo because the new bootstrap icons use position relative, and our popovers have a default zIndex of 0 (logic from WidgetStack). When the user needs a higher zIndex as in our demo, he just needs to specify it when creating the datepicker, like we do in http://alloyui.com/examples/datepicker/.
So instead of changing the code I just changed the demo to do this instead. Let me know if that's ok.