AUI-1257 Fix text input sizes on form builder for bootstrap 3
The .input-small, .input-medium and .input-large css classes don't exist anymore for bootstrap 3. Bootstrap says that input width should now be changed by using their grid css classes, like col-sm-2. This class needs to be in a wrapper div instead of directly on the input though, so this is changing our js to build the text input elements with that in mind.